Put Menu class in a different module
