+++ app.rb
@@ -3,6 +3,7 @@ require './rental'
 require './person'
 require './student'
 require './teacher'
+require './menu'
 
 class App
   attr_accessor :books, :rentals, :people, :student, :teacher
@@ -110,39 +111,4 @@ class App
   end
 
   private :push_person_to_list, :create_teacher, :create_student, :person_type_input
-end
-
-class Menu
-      def display_menu
-        puts "\nWelcome to School Library App!"
-        puts "\nPlease choose an option by entering a number:"
-        puts '1 - List all books'
-        puts '2 - List all people'
-        puts '3 - Create a person'
-        puts '4 - Create a book'
-        puts '5 - Create a rental'
-        puts '6 - List all rentals for a given person id'
-        puts '7 - Exit'
-      end
-    
-      def process_option(option, app)
-        menu = {
-          1 => :list_books,
-          2 => :list_people,
-          3 => :create_person,
-          4 => :create_book,
-          5 => :create_rental,
-          6 => :list_rentals_for_person,
-          7 => :exit
-        }
-      
-        if menu.key?(option)
-          app.send(menu[option])
-          return true if menu[option] == :exit
-        else
-          puts 'Invalid option.'
-        end
-      
-        false
-      end
 end
